fix: упрощает условия столкновений
