+++ packages/client/src/game/model/CollisionController/index.ts
@@ -1,0 +1,1 @@
+export { CollisionController } from './CollisionController'
