--- packages/client/src/game/model/СollisionController/index.ts
@@ -1,1 +1,0 @@
-export { СollisionController } from './СollisionController'
